Update comment on distinct() method in Stream.java
The comment describing the distinct() method on a stream mentioned that it determines uniqueness by using the equals() method.  The implementation of this method however, adds elements to a set and returns the result of the set.  In so doing, the distinct() method actually determines uniqueness as a result of the hashCode() method, rather than the equals() method.